refactor fixed_vector

+++ oneflow/core/register/dense_shape_view.h
@@ -8,7 +8,7 @@ namespace oneflow {
 class ShapeProto;
 class Shape;
 class PodPtr;
-template<typename T, long long kMaxSize>
+template<typename T, int kMaxSize>
 class fixed_vector;
 
 class DenseShapeView final {
